refactor: Split process child module
